requirements.txt
PyYAML==5.3.1
Quart==0.12.0